Create slack in pedal range.
Let the lowest 1/6th of the pedal range map onto 0. This prevents the
positive feedback loop where one hard kick on the pedal pushes the 
range down, so that the user needs to kick even harder to reach the 
lowest value again, which in turn pushes the range down, etc, etc.